scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.44.0
pymongo[srv]==4.7.3
python-dotenv==1.0.1
groq==0.9.0
pillow==10.3.0
tweepy==4.14.0
python-telegram-bot==21.3
brevo-python==3.1.0
lxml==5.2.2
cloudinary==1.40.0
algoliasearch==4.2.0
